% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_timepoint}
\alias{parse_timepoint}
\title{parse a timepoint into a long data.frame}
\usage{
parse_timepoint(timepoint)
}
\arguments{
\item{timepoint}{a timepoint}
}
\description{
parse a timepoint into a long data.frame
}
