% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldclimate.R
\name{fc_ping}
\alias{fc_ping}
\title{Ping fieldclimate API}
\usage{
fc_ping(timeout = 2)
}
\arguments{
\item{timeout}{number of seconds to wait for a response before giving up.}
}
\description{
Ping fieldclimate API
}
\examples{
\dontrun{
fc_ping()
}
}
